Fix en parseo de archivo, func lista. Funcion buscarInfractores

+++ C-GOD-MODE/Ejercicio4/Tp3-Ej4/info.h
@@ -6,7 +6,7 @@ typedef struct {
     char *patente;
     char *nombre_titular;
     int cantidad_multas;
-    double monto_total;
+    float monto_total;
 } t_dato;
